BUG: Exception during shutdown
The first reason was because Switcher.end_gpio() didn't check that we are running on Linux, and so issued GPIO.cleanup() on Windows.

The second reason was because if the can adapter open fails, we call MainWindow.shutdown, and this called Switcher.end_gpio(), and at this point we don't have a Switcher object, so I've added a check.