BUG: test data timing incorrect
The frequency of processing of each test data record was based on a rolling timer that accumulated the time offset values at the start of each record.  The problem was that, depending on the power of the hardware, the timings would drift due to accumlation of slight timing errors.  This code version sets a threshold epoch time for each record; once the threshold is crossed the record is generated.